[Feature] ContributionClaim And FeeWithdrawal (#3)
### Description
#### Claim Contribution Function:

The `claimContribution` function allows contributors to claim their contributions for a specific project. It performs necessary validations, updates contribution records, and transfers the claimed amount to the contributor.

#### Withdraw Platform Fee Function:
The `withdrawPlatformFee` function enables the contract owner to withdraw platform fees. It validates the currency type, checks fee availability, transfers the fees to the owner, and emits an event to indicate the withdrawal.

These changes enhance functionality and maintainability for contribution claiming and platform fee withdrawal.